requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose
psycopg2-binary
pydantic-settings
email-validator